WIP: Add basic landfall analysis figure
The `plotLandfallAnalysis.py` script reads in a collection of files that contain landfall rates for preliminary hazard simulations (1000 year simulations) and compares across the model groups, RCPs and time periods. Still only displays the figures, no saving. Assumes the files are in a specific folder structure (though this could readily be modified depending on how the files are stored - it's simply changing the format string for the file path).